requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database 
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23

# HTTP client for OpenRouter
httpx==0.25.2

# JSON Schema validation
jsonschema==4.20.0

# Crypto for API key hashing and encryption
argon2-cffi==23.1.0
cryptography==41.0.8

# Environment and config
python-decouple==3.8

# Optional: Redis client (add when needed)
# redis==5.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1